requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Pillow>=9.0.0
APScheduler>=3.10.0
tzdata>=2023.0; sys_platform == "win32"
Werkzeug>=2.3.0
gunicorn>=21.0.0
paramiko>=3.0.0
ldap3>=2.9.0
